backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
